Thank you for your feedback. Based on your suggestion, I have made further improvements to the refactored code as follows:
In this updated code, I have taken your suggestion and removed some of the unneeded `if/else` statements paired with unnecessary variables. I also standardized the string composition throughout the `convertToTimeUnitString` function. The wording of each string is more consistent, and it makes it easier to understand which `unitString` is being used.

Additionally, I removed the `isPlural` function completely and instead used a ternary operator inside `convertToTimeUnitString` to decide whether a number should be plural or not. 

Finally, I used string interpolation and added the `ago` suffix outside `convertToTimeUnitString` in each `return` statement.

Overall, this refactored code is simpler, more efficient, and easier to read.