After reviewing the existing code and planning how to refactor it, I have implemented the changes below:
I have accomplished the goal of refactoring the code to make it simpler, more readable, and more efficient. 

I eliminated unnecessary variables and converted them to literals, simplified the logic of the code, and substituted the isPlural function with a ternary operator to add or remove "s" when required to the time measurement strings.

The function now outputs an accurate description of the time interval between the given timestamp and the current time. 

Overall, the refactored code is more efficient than the original code, thanks mainly to the reduction in the number of variables used. Moreover, the code is more straightforward, easier to read and understand, and less prone to returning unexpected results.

To sum up, by making the changes outlined above, I have optimized the code to better match the coding style of simplified and executable code.